backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0
django-filter==23.1
djoser==2.1.0
drf_extra_fields==3.7.0
pillow==10.3.0
python-dotenv==1.0.1
